moved quicksearch controller method
