+++ bundles/AdminBundle/src/Controller/Searchadmin/SearchController.php
@@ -433,79 +433,6 @@ class SearchController extends AdminController
         return $query;
     }
 
-    /**
-     * @Route("/quicksearch", name="pimcore_admin_searchadmin_search_quicksearch", methods={"GET"})
-     *
-     * @param Request $request
-     * @param EventDispatcherInterface $eventDispatcher
-     *
-     * @return JsonResponse
-     */
-    public function quicksearchAction(Request $request, EventDispatcherInterface $eventDispatcher): JsonResponse
-    {
-        $query = $this->filterQueryParam($request->get('query', ''));
-        if (!preg_match('/[\+\-\*"]/', $query)) {
-            // check for a boolean operator (which was not filtered by filterQueryParam()),
-            // if present, do not add asterisk at the end of the query
-            $query = $query . '*';
-        }
-
-        $db = \Pimcore\Db::get();
-        $searcherList = new Data\Listing();
-
-        $conditionParts = [];
-
-        $conditionParts[] = $this->getPermittedPaths();
-
-        $matchCondition = '( MATCH (`data`,`properties`) AGAINST (' . $db->quote($query) . ' IN BOOLEAN MODE) )';
-        $conditionParts[] = '(' . $matchCondition . " AND `type` != 'folder') ";
-
-        $queryCondition = implode(' AND ', $conditionParts);
-
-        $searcherList->setCondition($queryCondition);
-        $searcherList->setLimit(50);
-        $searcherList->setOrderKey($matchCondition, false);
-        $searcherList->setOrder('DESC');
-
-        $beforeListLoadEvent = new GenericEvent($this, [
-            'list' => $searcherList,
-            'query' => $query,
-        ]);
-        $eventDispatcher->dispatch($beforeListLoadEvent, AdminEvents::QUICKSEARCH_LIST_BEFORE_LIST_LOAD);
-        $searcherList = $beforeListLoadEvent->getArgument('list');
-
-        $hits = $searcherList->load();
-
-        $elements = [];
-        foreach ($hits as $hit) {
-            $element = Element\Service::getElementById($hit->getId()->getType(), $hit->getId()->getId());
-            if ($element->isAllowed('list')) {
-                $data = [
-                    'id' => $element->getId(),
-                    'type' => $hit->getId()->getType(),
-                    'subtype' => $element->getType(),
-                    'className' => ($element instanceof DataObject\Concrete) ? $element->getClassName() : '',
-                    'fullpathList' => htmlspecialchars($this->shortenPath($element->getRealFullPath())),
-                ];
-
-                $this->addAdminStyle($element, ElementAdminStyleEvent::CONTEXT_SEARCH, $data);
-
-                $elements[] = $data;
-            }
-        }
-
-        $afterListLoadEvent = new GenericEvent($this, [
-            'list' => $elements,
-            'context' => $query,
-        ]);
-        $eventDispatcher->dispatch($afterListLoadEvent, AdminEvents::QUICKSEARCH_LIST_AFTER_LIST_LOAD);
-        $elements = $afterListLoadEvent->getArgument('list');
-
-        $result = ['data' => $elements, 'success' => true];
-
-        return $this->adminJson($result);
-    }
-
     protected function shortenPath(string $path): string
     {
         $parts = explode('/', trim($path, '/'));
