+++ bundles/SimpleBackendSearchBundle/public/js/pimcore/layout/quickSearch.js
@@ -27,7 +27,7 @@ Ext.define('pimcore.simpleBackendSearch.layout.quickSearch', {
                         var record = quickSearchStore.getAt(node.dataset.recordindex);
                         if (!record.get('preview')) {
                             Ext.Ajax.request({
-                                url: Routing.generate('pimcore_search_quicksearch_by_id#'),
+                                url: Routing.generate('pimcore_search_quicksearch_by_id'),
                                 method: 'GET',
                                 params: {
                                     "id": record.get('id'),
@@ -121,7 +121,7 @@ Ext.define('pimcore.simpleBackendSearch.layout.quickSearch', {
             this.quicksearchStore = new Ext.data.Store({
                 proxy: {
                     type: 'ajax',
-                    url: Routing.generate('pimcore_admin_searchadmin_search_quicksearch'),
+                    url: Routing.generate('pimcore_search_quicksearch'),
                     reader: {
                         type: 'json',
                         rootProperty: 'data'
